# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Testing ",
        "description": "Testing",
        "contact": {
            "email": "pikiha52@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api/v1",
            "description": "Testing"
        }
    ],
    "paths": {
        "/auth/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Signin",
                "description": "Api for",
                "operationId": "index",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "mrjohn@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "true"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email atau Password Anda salah"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/candidate/view": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Candidate",
                "description": "List all candidate",
                "operationId": "candidate",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "true"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/candidate/add": {
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Candidate",
                "description": "List all candidate",
                "operationId": "candidateAdd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "education": {
                                        "description": "education",
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "description": "birthday",
                                        "type": "string"
                                    },
                                    "experience": {
                                        "description": "experience",
                                        "type": "string"
                                    },
                                    "last_position": {
                                        "description": "last_position",
                                        "type": "string"
                                    },
                                    "applied_position": {
                                        "description": "applied_position",
                                        "type": "string"
                                    },
                                    "skils": {
                                        "description": "skils",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "phone",
                                        "type": "string"
                                    },
                                    "resume": {
                                        "description": "resume",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "true"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/candidate/read/{id}": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Candidate",
                "description": "List one candidate by id",
                "operationId": "candidateRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "true"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/candidate/edit/{id}": {
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Candidate",
                "description": "Edit candidate",
                "operationId": "candidateEdit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "education": {
                                        "description": "education",
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "description": "birthday",
                                        "type": "string"
                                    },
                                    "experience": {
                                        "description": "experience",
                                        "type": "string"
                                    },
                                    "last_position": {
                                        "description": "last_position",
                                        "type": "string"
                                    },
                                    "applied_position": {
                                        "description": "applied_position",
                                        "type": "string"
                                    },
                                    "skils": {
                                        "description": "skils",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "phone",
                                        "type": "string"
                                    },
                                    "resume": {
                                        "description": "resume",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "true"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/candidate/delete/{id}": {
            "delete": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Candidate",
                "description": "Delete one candidate",
                "operationId": "candidateDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "true"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}